+++ packages/backend-common/src/reading/GitlabUrlReader.test.ts
@@ -20,11 +20,16 @@ import fs from 'fs-extra';
 import mockFs from 'mock-fs';
 import { rest } from 'msw';
 import { setupServer } from 'msw/node';
+import os from 'os';
 import path from 'path';
 import { getVoidLogger } from '../logging';
 import { GitlabUrlReader } from './GitlabUrlReader';
 import { ReadTreeResponseFactory } from './tree';
 import { NotModifiedError, NotFoundError } from '../errors';
+import {
+  GitLabIntegration,
+  readGitLabIntegrationConfig,
+} from '@backstage/integration';
 
 const logger = getVoidLogger();
 
@@ -33,22 +38,44 @@ const treeResponseFactory = ReadTreeResponseFactory.create({
 });
 
 const gitlabProcessor = new GitlabUrlReader(
-  {
-    host: 'gitlab.com',
-    apiBaseUrl: 'https://gitlab.com/api/v4',
-  },
+  new GitLabIntegration(
+    readGitLabIntegrationConfig(
+      new ConfigReader({
+        host: 'gitlab.com',
+        apiBaseUrl: 'https://gitlab.com/api/v4',
+        baseUrl: 'https://gitlab.com',
+      }),
+    ),
+  ),
   { treeResponseFactory },
 );
 
 const hostedGitlabProcessor = new GitlabUrlReader(
-  {
-    host: 'gitlab.mycompany.com',
-    apiBaseUrl: 'https://gitlab.mycompany.com/api/v4',
-  },
+  new GitLabIntegration(
+    readGitLabIntegrationConfig(
+      new ConfigReader({
+        host: 'gitlab.mycompany.com',
+        apiBaseUrl: 'https://gitlab.mycompany.com/api/v4',
+        baseUrl: 'https://gitlab.mycompany.com',
+      }),
+    ),
+  ),
   { treeResponseFactory },
 );
 
+const tmpDir = os.platform() === 'win32' ? 'C:\\tmp' : '/tmp';
+
 describe('GitlabUrlReader', () => {
+  beforeEach(() => {
+    mockFs({
+      [tmpDir]: mockFs.directory(),
+    });
+  });
+
+  afterEach(() => {
+    mockFs.restore();
+  });
+
   const worker = setupServer();
   msw.setupDefaultHandlers(worker);
 
@@ -154,16 +181,6 @@ describe('GitlabUrlReader', () => {
   });
 
   describe('readTree', () => {
-    beforeEach(() => {
-      mockFs({
-        '/tmp': mockFs.directory(),
-      });
-    });
-
-    afterEach(() => {
-      mockFs.restore();
-    });
-
     const archiveBuffer = fs.readFileSync(
       path.resolve('src', 'reading', '__fixtures__', 'gitlab-archive.zip'),
     );
@@ -379,17 +396,81 @@ describe('GitlabUrlReader', () => {
       };
       await expect(fnGithub).rejects.toThrow(NotFoundError);
     });
+  });
 
-    it('should throw error when apiBaseUrl is missing', () => {
-      expect(() => {
-        /* eslint-disable no-new */
-        new GitlabUrlReader(
-          {
-            host: 'gitlab.mycompany.com',
-          },
-          { treeResponseFactory },
-        );
-      }).toThrowError('must configure an explicit apiBaseUrl');
+  describe('search', () => {
+    const archiveBuffer = fs.readFileSync(
+      path.resolve('src', 'reading', '__fixtures__', 'gitlab-archive.zip'),
+    );
+
+    const projectGitlabApiResponse = {
+      id: 11111111,
+      default_branch: 'main',
+    };
+
+    const branchGitlabApiResponse = {
+      commit: {
+        id: 'sha123abc',
+      },
+    };
+
+    beforeEach(() => {
+      worker.use(
+        rest.get(
+          'https://gitlab.com/api/v4/projects/backstage%2Fmock/repository/archive.zip?sha=main',
+          (_, res, ctx) =>
+            res(
+              ctx.status(200),
+              ctx.set('Content-Type', 'application/zip'),
+              ctx.set(
+                'content-disposition',
+                'attachment; filename="mock-main-sha123abc.zip"',
+              ),
+              ctx.body(archiveBuffer),
+            ),
+        ),
+        rest.get(
+          'https://gitlab.com/api/v4/projects/backstage%2Fmock',
+          (_, res, ctx) =>
+            res(
+              ctx.status(200),
+              ctx.set('Content-Type', 'application/json'),
+              ctx.json(projectGitlabApiResponse),
+            ),
+        ),
+        rest.get(
+          'https://gitlab.com/api/v4/projects/backstage%2Fmock/repository/branches/main',
+          (_, res, ctx) =>
+            res(
+              ctx.status(200),
+              ctx.set('Content-Type', 'application/json'),
+              ctx.json(branchGitlabApiResponse),
+            ),
+        ),
+      );
+    });
+
+    it('works for the naive case', async () => {
+      const result = await gitlabProcessor.search(
+        'https://gitlab.com/backstage/mock/tree/main/**/index.*',
+      );
+      expect(result.etag).toBe('sha123abc');
+      expect(result.files.length).toBe(1);
+      expect(result.files[0].url).toBe(
+        'https://gitlab.com/backstage/mock/tree/main/docs/index.md',
+      );
+      await expect(result.files[0].content()).resolves.toEqual(
+        Buffer.from('# Test\n'),
+      );
+    });
+
+    it('throws NotModifiedError when same etag', async () => {
+      await expect(
+        gitlabProcessor.search(
+          'https://gitlab.com/backstage/mock/tree/main/**/index.*',
+          { etag: 'sha123abc' },
+        ),
+      ).rejects.toThrow(NotModifiedError);
     });
   });
 });