+++ packages/core-api/src/icons/types.ts
@@ -17,6 +17,8 @@
 import { ComponentType } from 'react';
 import { SvgIconProps } from '@material-ui/core';
 
+export type SystemIconKey = 'user' | 'group' | 'dashboard' | 'help';
+
 export type IconComponent = ComponentType<SvgIconProps>;
-export type SystemIconKey = 'user' | 'group';
-export type SystemIcons = { [key in SystemIconKey]: IconComponent };
+export type IconKey = SystemIconKey | string;
+export type IconComponentMap = { [key in IconKey]: IconComponent };