+++ plugins/auth-backend/config.d.ts
@@ -58,7 +58,13 @@ export interface Config {
         development: { [key: string]: string };
       };
       oauth2?: {
-        development: { [key: string]: string };
+        development: {
+          clientId: string;
+          clientSecret: string;
+          authorizationUrl: string;
+          tokenUrl: string;
+          scope?: string;
+        };
       };
       oidc?: {
         development: { [key: string]: string };