+++ plugins/auth-backend/src/identity/MemoryKeyStore.ts
@@ -14,18 +14,15 @@
  * limitations under the License.
  */
 
-import { utc } from 'moment';
 import { KeyStore, AnyJWK, StoredKey } from './types';
+import { DateTime } from 'luxon';
 
 export class MemoryKeyStore implements KeyStore {
-  private readonly keys = new Map<
-    string,
-    { createdAt: moment.Moment; key: string }
-  >();
+  private readonly keys = new Map<string, { createdAt: Date; key: string }>();
 
   async addKey(key: AnyJWK): Promise<void> {
     this.keys.set(key.kid, {
-      createdAt: utc(),
+      createdAt: DateTime.utc().toJSDate(),
       key: JSON.stringify(key),
     });
   }