+++ plugins/catalog-backend/src/ingestion/processors/github/github.ts
@@ -27,6 +27,7 @@ export type Organization = {
   membersWithRole?: Connection<User>;
   team?: Team;
   teams?: Connection<Team>;
+  repositories?: Connection<Repository>;
 };
 
 export type PageInfo = {
@@ -50,6 +51,11 @@ export type Team = {
   avatarUrl?: string;
   parentTeam?: Team;
   members: Connection<User>;
+};
+
+export type Repository = {
+  name: string;
+  url: string;
 };
 
 export type Connection<T> = {
@@ -214,6 +220,39 @@ export async function getOrganizationTeams(
   );
 
   return { groups, groupMemberUsers };
+}
+
+export async function getOrganizationRepositories(
+  client: typeof graphql,
+  org: string,
+): Promise<{ repositories: Repository[] }> {
+  const query = `
+  query repositories($org: String!, $cursor: String) {
+    organization(login: $org) {
+      name
+      repositories(first: 100, after: $cursor) {
+        nodes {
+          name
+          url
+        }
+        pageInfo {
+          hasNextPage
+          endCursor
+        }
+      }
+    }
+  }
+  `;
+
+  const repositories = await queryWithPaging(
+    client,
+    query,
+    r => r.organization?.repositories,
+    x => x,
+    { org },
+  );
+
+  return { repositories };
 }
 
 /**