+++ plugins/catalog-react/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-catalog-react
 
+## 0.0.3
+
+### Patch Changes
+
+- 19d354c78: Make `EntityRefLink` a `React.forwardRef` in order to use it as root component in other components like `ListItem`.
+- Updated dependencies [6ed2b47d6]
+- Updated dependencies [72b96e880]
+- Updated dependencies [b51ee6ece]
+  - @backstage/catalog-client@0.3.6
+  - @backstage/core@0.6.1
+
 ## 0.0.2
 
 ### Patch Changes