+++ plugins/techdocs-backend/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @backstage/plugin-techdocs-backend
 
+## 0.6.0
+
+### Minor Changes
+
+- 08142b256: URL Preparer will now use proper etag based caching introduced in https://github.com/backstage/backstage/pull/4120. Previously, builds used to be cached for 30 minutes.
+
+### Patch Changes
+
+- 08142b256: TechDocs will throw warning in backend logs when legacy git preparer or dir preparer is used to preparer docs. Migrate to URL Preparer by updating `backstage.io/techdocs-ref` annotation to be prefixed with `url:`.
+  Detailed docs are here https://backstage.io/docs/features/techdocs/how-to-guides#how-to-use-url-reader-in-techdocs-prepare-step
+  See benefits and reason for doing so https://github.com/backstage/backstage/issues/4409
+- Updated dependencies [77ad0003a]
+- Updated dependencies [ffffea8e6]
+- Updated dependencies [82b2c11b6]
+- Updated dependencies [965e200c6]
+- Updated dependencies [5a5163519]
+- Updated dependencies [08142b256]
+- Updated dependencies [08142b256]
+  - @backstage/techdocs-common@0.4.0
+  - @backstage/backend-common@0.5.3
+
 ## 0.5.5
 
 ### Patch Changes